+++ modules/AWSPowerShell/Cmdlets/DirectoryService/AWS.Tools.DirectoryService.Completers.psm1
@@ -94,7 +94,14 @@ $DS_Completers = {
             ($_ -eq "Get-DSClientAuthenticationSetting/Type")
         }
         {
-            $v = "SmartCard"
+            $v = "SmartCard","SmartCardOrPassword"
+            break
+        }
+
+        # Amazon.DirectoryService.DirectoryConfigurationStatus
+        "Get-DSSetting/Status"
+        {
+            $v = "Default","Failed","Requested","Updated","Updating"
             break
         }
 
@@ -192,6 +199,7 @@ $DS_map = @{
     "ShareMethod"=@("Enable-DSDirectoryShare")
     "ShareTarget_Type"=@("Enable-DSDirectoryShare")
     "Size"=@("Connect-DSDirectory","New-DSDirectory")
+    "Status"=@("Get-DSSetting")
     "TrustDirection"=@("New-DSTrust")
     "TrustType"=@("New-DSTrust")
     "Type"=@("Disable-DSClientAuthentication","Disable-DSLDAPS","Enable-DSClientAuthentication","Enable-DSLDAPS","Get-DSClientAuthenticationSetting","Get-DSLDAPSSetting","Register-DSCertificate")
@@ -277,6 +285,7 @@ $DS_SelectMap = @{
                "Get-DSEventTopic",
                "Get-DSLDAPSSetting",
                "Get-DSRegion",
+               "Get-DSSetting",
                "Get-DSSharedDirectory",
                "Get-DSSnapshot",
                "Get-DSTrust",
@@ -309,6 +318,7 @@ $DS_SelectMap = @{
                "Update-DSConditionalForwarder",
                "Set-DSDomainControllerCount",
                "Update-DSRadius",
+               "Update-DSSetting",
                "Update-DSTrust",
                "Approve-DSTrust")
 }